Re-implmented Most Debuffs and Combat Effects
Reimplemented most combat effects based on the new logic.

+++ Assets/scripts/GameManagers/BuffsDebuffs/Debuff.cs
@@ -7,7 +7,7 @@ public abstract class Debuff : ScriptableObject, CardEffectInterface
 {
     public string debuffName;
     public int turnDuration;
-    public float intensity;
+    public int intensity;
     public CharacterInstance target;
 
     public void ResolveEffect(Card card, CharacterInstance character, int cost, CombatManager cManager)
